+++ ChangeLog.md
@@ -1,4 +1,4 @@
-# pending
+# v1.8.0
 * Critical rolls on initiative affect the degree of success of the sneak evaluation
 * Added a condition handler setting, allowing a choice of how to handle setting the `hidden`, `undetected`, and `unnoticed` settings at your table.